Move new code page notification code to a more appropriate place
