+++ src/dos/dos_code_page.cpp
@@ -29,6 +29,7 @@
 #include <vector>
 
 #include "../ints/int10.h"
+#include "autoexec.h"
 #include "bios.h"
 #include "checks.h"
 #include "dos_locale.h"
@@ -43,6 +44,17 @@ CHECK_NARROWING();
 // References:
 // - https://www.seasip.info/DOS/CPI/cpi.html
 // - http://kbd-project.org/docs/font-formats/font-formats-3.html
+
+extern void DOS_UpdateCurrentProgramName();
+
+static void notify_code_page_changed()
+{
+	// Recreate various information to match new code page
+	MSG_NotifyNewCodePage();
+	DOS_UpdateCurrentProgramName();
+	DOS_RepopulateCountryInfo();
+	AUTOEXEC_NotifyNewCodePage();
+}
 
 // ***************************************************************************
 // Constants
@@ -1989,6 +2001,7 @@ static KeyboardLayoutResult load_custom_screen_font(const uint16_t code_page,
 	        code_page,
 	        canonical_file_name.c_str());
 
+	notify_code_page_changed();
 	return KeyboardLayoutResult::OK;
 }
 
@@ -1996,7 +2009,7 @@ static KeyboardLayoutResult load_bundled_screen_font(const uint16_t code_page)
 {
 	if (dos.loaded_codepage == code_page &&
 	    dos.screen_font_type == ScreenFontType::Bundled) {
-		// Already loaded - skip
+		// Already loaded - skip loading and notifying other subsystems
 		return KeyboardLayoutResult::OK;
 	}
 
@@ -2014,14 +2027,15 @@ static KeyboardLayoutResult load_bundled_screen_font(const uint16_t code_page)
 	        code_page,
 	        DOS_GetCodePageDescriptionForLog(code_page).c_str());
 
+	notify_code_page_changed();
 	return KeyboardLayoutResult::OK;
 }
 
 static void load_default_screen_font()
 {
-	if (dos.loaded_codepage == DefaultCodePage &&
+	if (dos.loaded_codepage  == DefaultCodePage &&
 	    dos.screen_font_type == ScreenFontType::Rom) {
-		// Already loaded - skip
+		// Already loaded - skip loading and notifying other subsystems
 		return;
 	}
 
@@ -2034,6 +2048,8 @@ static void load_default_screen_font()
 	dos.screen_font_type = ScreenFontType::Rom;
 
 	LOG_MSG("LOCALE: Loaded code page %d (ROM font)", dos.loaded_codepage);
+
+	notify_code_page_changed();
 }
 
 // ***************************************************************************
